Made Graph class and clique finding code templated

+++ ORBITER/src/lib/foundations/graph_theory/Clique/main.cpp
@@ -6,9 +6,12 @@ using std::endl;
 
 int main () {
 
+//	typedef uint32_t _vertex_lbl_type_;
+//	typedef uint16_t _vertex_color_type_;
+
 	Graph G (5); // Generate the graph
 
-	std::vector<std::vector<size_t>> v; // Vector containing solutions
+	std::vector<std::vector<uint32_t>> v; // Vector containing solutions
 
 	RainbowClique::find_cliques(G, v);
 }
